Sửa lỗi khi cập nhật sản phẩm

+++ server/routes/api/sanPham.js
@@ -388,11 +388,12 @@ router.post('/', validator(sanPhamSchema), async (req, res) => {
 router.put('/', validator(sanPhamUpdateSchema), async (req, res) => {
     try {
         var collectionSanPham = db.collection('SanPham');
+        var loai_san_pham = (await db.collection('LoaiSanPham').
+            doc(req.body.ma_loai_san_pham || '-').get())?.data()?.loai_san_pham;
         var updateSanPham = {
             ten_san_pham: req.body.ten_san_pham,
             ma_loai_san_pham: req.body.ma_loai_san_pham,
-            loai_san_pham: (await db.collection('LoaiSanPham').
-                doc(req.body.ma_loai_san_pham).get()).data().loai_san_pham,
+            loai_san_pham: loai_san_pham,
             gia_tien: req.body.gia_tien,
             so_luong: req.body.so_luong,
             mo_ta: req.body.mo_ta,
@@ -405,11 +406,11 @@ router.put('/', validator(sanPhamUpdateSchema), async (req, res) => {
         var result = await collectionSanPham.doc(req.body.ma_san_pham).update(updateSanPham);
         updateSanPham.ma_san_pham = result.id;
         var fileURL_arr = [];
-        console.log(req.body.delete_file.split(','))
-        for (var deleteURL of req.body.delete_file.split(',')) {
-            firebaseApp.storage().refFromURL(deleteURL).delete();
-            await collectionSanPham.doc(req.body.ma_san_pham).update({ file: FieldValue.arrayRemove(deleteURL) });
-        }
+        if (req.body.delete_file != null)
+            for (var deleteURL of req.body.delete_file.split(',')) {
+                firebaseApp.storage().refFromURL(deleteURL).delete();
+                await collectionSanPham.doc(req.body.ma_san_pham).update({ file: FieldValue.arrayRemove(deleteURL) });
+            }
         if (req.files?.file)
             for (const file of req.files?.file) {
                 var snapshot = await firebaseApp.storage()
@@ -420,7 +421,7 @@ router.put('/', validator(sanPhamUpdateSchema), async (req, res) => {
                 console.log(fileURL);
                 fileURL_arr.push(fileURL);
             }
-        updateSanPham.delete_file = req.body.delete_file.split(',');
+        updateSanPham.delete_file = req.body.delete_file?.split(',');
         updateSanPham.new_file = fileURL_arr;
         return res.json({ success: true, data: updateSanPham });
     }
